PAC.Vidly.WebApi/Controllers/Movies/MovieController.cs
using Microsoft.AspNetCore.Mvc;
using PAC.Vidly.WebApi.Controllers.Movies.Models;
using PAC.Vidly.WebApi.Filters;
using PAC.Vidly.WebApi.Services.Movies;
using PAC.Vidly.WebApi.Services.Movies.Entities;
using PAC.Vidly.WebApi.Services.Users.Entities;

namespace PAC.Vidly.WebApi.Controllers.Movies
{
    [ApiController]
    [Route("movie")]
    [AuthenticationFilter]
    public sealed class MovieController : ControllerBase
    {
        private readonly IMovieService _movieService;

        public MovieController(IMovieService movieService)
        {
            _movieService = movieService;
        }

        [HttpPost]
        public Movie Create(Movie? request)
        {
            if (request == null)
            {
                throw new ArgumentNullException(nameof(request));
            }

            var userLogged = GetUserLogged();

            var movie = _movieService.Create(request, userLogged.Id);
            return movie;
        }

        private User GetUserLogged()
        {
            var userLogged = HttpContext.Items[Items.UserLogged] as User;
            return userLogged;
        }

        [HttpGet]
        public List<MovieBasicInfoResponse> GetAll()
        {
            var movies = _movieService.GetAll().Select(m => 
                new MovieBasicInfoResponse(m)).ToList();

            return movies;
        }
    }
}
